MATLAB
According to the Nyquist Criterion, in order to accurately generate the f (t) signal, it is necessary to make a sampling of at least twice the highest frequency component of f (t). One of the approaches used to accomplish this is the sampling of the signal by selecting the sampling frequency 5 times larger than the highest frequency component.